rag-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
mysql-connector-python==8.3.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LlamaIndex and Vector Store
llama-index==0.10.11
llama-index-vector-stores-chroma==0.1.4
llama-index-embeddings-google==0.1.3
llama-index-llms-gemini==0.1.6
chromadb==0.4.22

# Google Generative AI
google-generativeai==0.3.2

# Utilities
requests==2.31.0
